requirements.txt
accelerate
pydub
nnAudio
PyYAML
transformers==4.53.3
hydra-core
tensorboard
slider @ git+https://github.com/OliBomby/slider.git@gedagedigedagedaoh
torch_tb_profiler
wandb
matplotlib
pandas
pyarrow
einops
lightning
pywebview[qt]
pyqtwebengine
flask
audioop-lts; python_version>='3.13'
rich
torchaudio